声鉴

+++ app/controllers/m/UsersController.php
@@ -285,17 +285,19 @@ class UsersController extends BaseController
                     $img_path = str_replace('../../..', '', $new_file);
                     $img_files[] = $img_path;
                     $res = \Albums::uploadImage($user, $img_files);
-                    if ($res) {
-                        if (file_exists($source_filename)) {
-                            unlink($source_filename);
+                    $image_result = $user->saveImageToPhone();
+                    if($res){
+                        if ($image_result) {
+                            if (file_exists($source_filename)) {
+                                unlink($source_filename);
+                            }
+                            return $this->renderJSON(ERROR_CODE_SUCCESS, '图片上传成功');
+                        } else {
+                            return $this->renderJSON(ERROR_CODE_FAIL, '图片上传失败');
                         }
-                        return $this->renderJSON(ERROR_CODE_SUCCESS, '图片上传成功');
-                    } else {
-                        return $this->renderJSON(ERROR_CODE_FAIL, '图片上传失败');
                     }
                 } else {
                     return $this->renderJSON(ERROR_CODE_FAIL, '图片生成失败');
-
                 }
             } else {
                 //文件类型错误
